hermes-web/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.11.7
python-dotenv==1.0.0
PyJWT==2.10.1
alembic==1.12.1
psycopg[binary]==3.2.3
pgvector==0.2.5
